requirements.txt
flask
openai
pandasql
langchain
python-dotenv
langchain-community
markdown
PyPDF2